% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixFunctions.R
\name{addZeroDiagonal}
\alias{addZeroDiagonal}
\title{Adding zero diagonal to a matrix}
\usage{
addZeroDiagonal(M)
}
\arguments{
\item{M}{a matrix to which you need to add zero diagonal}
}
\description{
Function adds zero diagonal to \code{(p-1)} by \code{p} matrix and returns \code{p} by \code{p} matrix
}
\keyword{internal}
