% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLambdaCV.R
\name{estimateLambdaCV}
\alias{estimateLambdaCV}
\title{Estimation of optimal lambda parameter for jewel method based on cross-validation}
\usage{
estimateLambdaCV(X, lambda, k_folds = 5, verbose = TRUE, makePlot = TRUE)
}
\arguments{
\item{X}{list of \code{K} data matrices of size \code{n_k} by \code{p} (\code{n_k} can be different for each class)}

\item{lambda}{vector of grid lambda parameters over which cross-validation is performed}

\item{k_folds}{number of folds in which data is divided. The default value is 5.}

\item{verbose}{If verbose = FALSE, tracing information printing is disabled. The default value is TRUE.}

\item{makePlot}{If makePlot = FALSE, plotting of CV error is disabled. The default value is TRUE.}
}
\value{
The following list is returned
\itemize{
  \item lambda_opt - a number, optimal value of regularization parameter according to cross-validation procedure
  \item CV_err - a vector of cross-validation errors for each element of input vector \code{lambda}
}
}
\description{
Given a grid of lambda parameters, function performs cross-validation and finds the optimal one. Parallelization over folds and warm start are implemented.
}
