% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate a ''true'' graph and corresponding data matrices}
\usage{
generateData(
  K,
  n,
  p,
  power = 1,
  m = 1,
  c = 0.2,
  d = 0.8,
  makePlot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{K}{Number of data matrices}

\item{n}{Vector of the sample sizes for each desired set of data matrices. Can be a vector of one element.}

\item{p}{Number of nodes in the true graph.}

\item{power}{Power of preferential attachment for Barabasi-Albert algorithm for generation of scale-free graph.}

\item{m}{Number of edges to add in each time step of Barabasi-Albert algorithm for generation of scale-free graph.}

\item{c}{Entries in precision matrices are generated from the uniform distribution on the interval \code{[-d, -c] + [c, d]}. The default value is \code{c = 0.2}.}

\item{d}{Entries in precision matrices are generated from the uniform distribution on the interval \code{[-d, -c] + [c, d]}. The default value is \code{d = 0.8}.}

\item{makePlot}{If makePlot = FALSE, plotting of the generated true graph is disabled. The default value is TRUE.}

\item{verbose}{If verbose = FALSE, tracing information printing is disabled. The default value is TRUE.}
}
\value{
The following list is returned
 \itemize{
   \item trueGraph - sparse adjacency matrix of the true graph
   \item data - list of lists, \code{K} data matrices of the size \code{n_l} by \code{p} for each sample size element \code{l} of the input vector \code{n}
   \item Sigma - list of \code{K} covariance matrices of the size \code{p} by \code{p}
}
}
\description{
Function generates a scale-free graph with \code{p} vertices, \code{K} corresponding precision and covariance matrices, all of dize \code{p} by \code{p}
and then for each \code{l}-th element of vector \code{n} generates \code{K} data matrices of size \code{n_l} by \code{p}. For the same underlying graph we can generate several datasets with different sample sizes.
}
