% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLambdaGrid.R
\name{generateLambdaGrid}
\alias{generateLambdaGrid}
\title{Generate a sequence of lambda parameters}
\usage{
generateLambdaGrid(X, n = 50, eps = 0.1, scale = TRUE)
}
\arguments{
\item{X}{list of \code{K} data matrices of size \code{n_k} by {p} (\code{n_k} can be different for each class).}

\item{n}{desired number of parameters. The default value is 50.}

\item{eps}{lambda_min = lambda_max * eps. The default value is 0.1}

\item{scale}{if TRUE and data is scaled, than resulting grid is independent of X and goes from eps to 1 uniformly in log-scale. The default value is TRUE.}
}
\value{
vector of regularization parameters of length \code{n}
}
\description{
Function that generates a uniform in logarithmic space grid of tuning parameters. Lambda_max is max(\code{1 / (n - 1)} max(X^TX)),  lambda_min = lambda_max * eps.
}
