% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixFunctions.R
\name{removeDiagonal}
\alias{removeDiagonal}
\title{Removing diagonal from a matrix}
\usage{
removeDiagonal(M)
}
\arguments{
\item{M}{a matrix in which you need to remove a diagonal}
}
\description{
Function removes a diagonal from a square \code{p} by \code{p} matrix
and returns \code{(p-1)} by \code{p} matrix
}
\keyword{internal}
