% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLambdaBIC.R
\name{estimateLambdaBIC}
\alias{estimateLambdaBIC}
\title{Estimation of optimal lambda parameter for jewel method with Bayesian information criterion}
\usage{
estimateLambdaBIC(X, lambda)
}
\arguments{
\item{X}{list of \code{K} data matrices of size \code{n_k} by \code{p} (\code{n_k} can be different for each class)}

\item{lambda}{vector of grid lambda parameters for which function evaluates BIC}
}
\value{
The following list is returned
\itemize{
  \item lambda_opt - a number, optimal value of regularization parameter according to BIC procedure
  \item CV_err - a vector of cross-validation erros for each element of input vector \code{lambda}.
}
}
\description{
Given a grid of lambda parameters, function evaluates Baeysian information criterion for each element.
Optimal lambda is chosen as a BIC's minimum. Warm start is implemented.
}
