% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLambdaBIC.R
\name{estimateLambdaBIC}
\alias{estimateLambdaBIC}
\title{Estimate the optimal regularization parameter for jewel method with Bayesian information criterion}
\usage{
estimateLambdaBIC(X, lambda = NULL, tol = 1e-04, makePlot = TRUE)
}
\arguments{
\item{X}{a list of \code{K} numeric data matrices of \code{n_k} samples and \code{p} variables (\code{n_k} can be different for each matrix).}

\item{lambda}{an optional numeric vector of parameters for which the function evaluates BIC. Users are encouraged to tailor the grid to their specific data. If NULL, the default value is uniform in log sequence from 0.01 to 0.5.}

\item{tol}{an optional number, convergence threshold controlling the relative error between iterations. The smaller it is, more precise is BIC. The default value is 0.0001.}

\item{makePlot}{If makePlot = FALSE, plotting of BIC is disabled. The default value is TRUE.}
}
\value{
The following list is returned
\itemize{
  \item \code{lambda_opt} - a number, optimal value of regularization parameter according to BIC procedure;
  \item \code{BIC} - a numeric vector of BICs for each element of the grid of \code{lambda}.
}
}
\description{
Function evaluates Baeysian information criterion (BIC) for each element of the grid.
Optimal lambda is chosen as the one for which BIC's minimum is obtained. Warm start is implemented (matrices computed with the previous parameter are used as a starting point for the current element of the grid).
}
