% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generation of a scale-free graph and corresponding datasets using the graph as their Gaussian graphical model}
\usage{
generateData(
  p,
  K,
  n,
  power = 1,
  m = 1,
  a = 0.2,
  b = 0.8,
  makePlot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{p}{Number of nodes in the true graph}

\item{K}{Number of data matrices}

\item{n}{Vector of the sample sizes for each desired set of \code{K} data matrices. Can be a vector of one element if one wishes to obtain only one dataset of \code{K} matrices.}

\item{power}{Power of preferential attachment for Barabasi-Albert algorithm for generation of the scale-free graph. The defaut value is 1.}

\item{m}{Number of edges to add at each time step of Barabasi-Albert algorithm for generation of the scale-free graph. Resulting graph has \code{mp - (2m - 1)} edges. The default value is 1 and graph has \code{p-1} edges.}

\item{a}{Entries of precision matrices are sampled from the uniform distribution on the interval \code{[-b, -a] + [a, b]}. The default value is \code{a = 0.2}.}

\item{b}{Entries of precision matrices are sampled from the uniform distribution on the interval \code{[-b, -a] + [a, b]}. The default value is \code{b = 0.8}.}

\item{makePlot}{If makePlot = FALSE, plotting of the generated true graph is disabled. The default value is TRUE.}

\item{verbose}{If verbose = FALSE, tracing information printing is disabled. The default value is TRUE.}
}
\value{
The following list is returned
 \itemize{
   \item \code{trueGraph} - sparse adjacency matrix of the true graph
   \item \code{data} - list of lists, for each sample size (\code{l}-th element of the input vector \code{n}) one obtains \code{K} data matrices, each of the size \code{n_l} by \code{p}
   \item \code{Sigma} - list of \code{K} covariance matrices of the size \code{p} by \code{p}
}
}
\description{
Function generates a scale-free graph with \code{p} vertices and \code{K} corresponding precision and covariance matrices, all of the size \code{p} by \code{p}.
Then for each \code{l}-th element of vector \code{n} it generates \code{K} data matrices, each of the size \code{n_l} by \code{p}, 
i.e., for the same underlying graph we can generate several sets of \code{K} datasets with different sample sizes.
}
