% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLambdaGrid.R
\name{generateLambdaGrid}
\alias{generateLambdaGrid}
\title{Generation of the sequence of regularization parameters}
\usage{
generateLambdaGrid(X, n = 50, eps = 0.1, scale = TRUE)
}
\arguments{
\item{X}{list of \code{K} numeric data matrices of size \code{n_k} by {p} (\code{n_k} can be different for each matrix)}

\item{n}{desired number of parameters. The default value is 50.}

\item{eps}{lambda_min = lambda_max * eps. The default value is 0.1}

\item{scale}{if TRUE and hence data is scaled, then resulting grid is independent of \code{X} and goes from \code{eps} to 1 uniformly in log-scale. The default value is TRUE.}
}
\value{
\code{lambda} - vector of regularization parameters of length \code{n}
}
\description{
Function generates a uniform in logarithmic space grid of regularization parameters. \code{Lambda_max} is \code{max({1 / (n - 1)} max(X^TX))},  \code{lambda_min = lambda_max * eps}.
}
