% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jewel.R
\name{jewel}
\alias{jewel}
\title{Joint node-wise estimation of multiple Gaussian graphical models}
\usage{
jewel(X, lambda, Theta = NULL, reltol = 0.01, maxIter = 10000, verbose = TRUE)
}
\arguments{
\item{X}{List of \code{K} numeric data matrices of size \code{n_k} by \code{p} (\code{n_k} can be different for each class).}

\item{lambda}{Tuning parameter which controls the sparsity of the resulting graph - bigger it is, less edges you get.}

\item{Theta}{List of \code{K} starting regression coefficient matrices of size \code{p} by \code{p}. If not provided, initialized as all zeros.}

\item{reltol}{Convergence threshold controlling relative error between iterations. The default value is 0.01.}

\item{maxIter}{Maximum allowed number of iterations. The default  value is 10 000.}

\item{verbose}{If verbose = FALSE, tracing information printing is disabled. The default value is TRUE.}
}
\value{
The following list is returned
\itemize{
  \item EstAdjMat - adjacency matrix of the estimated graph
  \item Theta - list of \code{K} estimated covariance matrices of size \code{p} by \code{p}
  \item BIC - value of Bayesian information criterion
  \item residual - list of \code{K} resulting residuals X - X * Theta of size  \code{n} by \code{p}
}
}
\description{
Method for estimating the graph of conditional depndencies between the variables given multiple dataset (observations of variables under different conditiona or collected in distinct classes).
}
